// frontend/src/types/agent.ts
export interface AgentTemplate {
    _id: string;
    _key: string;
    name: string;
    description: string;
    category: string;
    startMessage: string;
    systemPrompt: string;
    tools: string[];
    models: string[];
    apps: string[];
    kb: string[];
    vectorDBs: string[];
    tags: string[];
    createdBy: string;
    createdAt: string;
    updatedAt: string;
    usageCount: number;
    rating: number;
    icon?: string;
    isDeleted?: boolean;
  }